[Tests] Make test daemon available to users
Moved the code of the test daemon to the new p2pclient.test_daemon
module. This enables users to reuse the daemon in their own tests.
